refactor(webui): share code-block icon vocabulary across renderers
Extract the langtag→lucide-icon mapping (and the label-HTML builder) into
codeBlockIcons.ts and reuse it in the chat renderer AND the non-chat
preview components (TabbedCodeBlock, MarkdownPreviewTabs), replacing their
emoji labels with the matching lucide icons.

Shares only the icon *vocabulary*, not chat-message *chrome* (collapse,
tool-label sidebar) — those stay in the chat renderer. Making preview
markdown bodies render as fully standard markdown is a separate follow-up.
